requirements.txt
langchain==0.1.6
langchain-google-genai==0.0.9
google-generativeai>=0.3.1
faiss-cpu==1.7.4
streamlit
python-dotenv==1.0.0
pypdf==4.0.1
tiktoken==0.5.2
langchain-community==0.0.19
pydantic==2.5.2
gitpython==3.1.41